use max withdrawals per payload from the spec config

+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigCapella.java
@@ -36,5 +36,5 @@ public interface SpecConfigCapella extends SpecConfigBellatrix {
   @Override
   Optional<SpecConfigCapella> toVersionCapella();
 
-  int getMaxWithdrawalsPerPayload();
+  UInt64 getMaxWithdrawalsPerPayload();
 }
